Renamed Email to Inbox and refactored related functions
User's are really getting an inbox so it makes more sense for this
type to be called Inbox. Lot's of renaming functions and vars inside
of functions to better fit this new name. Also added several comments
to exported types to satisfy gometalinter and changed the session key
to a const.