Moved wg.Add from lambdaCreateRouteAndUpdate to prevent data race
Race detector was having a field day with this. When using a db with very little
latency (i.e InMemory) it is possible that wg.Wait() could be called before the
goroutine has time to run wg.Add(1). To prevent this we are going to call it before
starting the goroutine.