Fix links to js pages not working
Trys to fix an ongoing issue. Because the iframe an email is
rendered in blocks js, any windows/tabs opened from that iframe
also blocks js. Therefore if the page a user wishes to navigate
to relies on js it almost always breaks the page. This allows
js inside the iframe, but not on the website, so XSS should be
prevented here. I've also added a note to try explain this to
users.